Update TPS2216 to use ResistorVoltageDivider
Split of into reference design and bare module
