+++ src/faebryk/library/TPS2116.py
@@ -2,7 +2,6 @@
 # SPDX-License-Identifier: MIT
 
 import logging
-from enum import Enum, auto
 
 import faebryk.library._F as F  # noqa: F401
 from faebryk.libs.library import L  # noqa: F401
@@ -17,58 +16,6 @@ class TPS2116(F.PowerMux):
     """
     2 to 1 1.6 V to 5.5 V, 2.5-A Low IQ Power Mux with Manual and Priority Switchover
     """
-
-    class Mode(Enum):
-        MANUAL = 0
-        """
-        Manually tie mode to an external power reference.
-        If select is above Vref (1V), power_in[0] is selected.
-        If select is below Vref, power_in[1] is selected.
-        """
-        PRIORITY = 1
-        """
-        This is the most automatic mode.
-        power_in[0] is selected by default, switchover only happens if power_in[0] is
-        lower than power_in[1].
-        """
-        SHUTDOWN = 2
-        """
-        Disables device.
-        """
-
-    class SwitchoverVoltage(Enum):
-        _5V = auto()
-        _3V3 = auto()
-        _1V8 = auto()
-        CUSTOM = auto()
-
-    @assert_once
-    def set_mode(self, mode: Mode, switchover_voltage: SwitchoverVoltage):
-        if mode == self.Mode.PRIORITY:
-            self.mode.set(on=True)
-            resistor_divider = self.select.add(
-                F.ResistorVoltageDivider()
-            )  # TODO move to reference design
-            self.power_in[0].connect(resistor_divider.power)
-            self.select.line.connect(resistor_divider.output.line)
-            if switchover_voltage != self.SwitchoverVoltage.CUSTOM:
-                resistor_divider.r_bottom.resistance.constrain_subset(
-                    L.Range.from_center_rel(5 * P.kohm, 0.01)
-                )
-            if switchover_voltage == self.SwitchoverVoltage._5V:
-                resistor_divider.r_top.resistance.constrain_subset(
-                    L.Range.from_center_rel(16.9 * P.kohm, 0.01)
-                )
-            elif switchover_voltage == self.SwitchoverVoltage._3V3:
-                resistor_divider.r_top.resistance.constrain_subset(
-                    L.Range.from_center_rel(9.53 * P.kohm, 0.01)
-                )
-            elif switchover_voltage == self.SwitchoverVoltage._1V8:
-                resistor_divider.r_top.resistance.constrain_subset(
-                    L.Range.from_center_rel(2.80 * P.kohm, 0.01)
-                )
-        else:
-            raise NotImplementedError(f"Mode {mode} not implemented")
 
     # ----------------------------------------
     #     modules, interfaces, parameters
